# Demonstrating surprising performance of "not hashing partition number" alone