--- a/fuzz-tests/Cargo.toml
+++ b/fuzz-tests/Cargo.toml
@@ -77,3 +77,5 @@ verbose = []
 fuzzer = [ "simple-fuzzer" ]
 
 rocksdb = []
+parnum_hashed = []
+parnum_not_hashed = []